refactor: 파일 구조 변경 계층형 디렉토리 구조로 변경
