--- src/main/java/org/programmers/springorder/customer/model/CustomerType.java
@@ -1,5 +1,0 @@
-package org.programmers.springorder.customer.model;
-
-public enum CustomerType {
-    NORMAL, BLACK
-}
