+++ src/main/java/org/programmers/springorder/dto/customer/CustomerResponseDto.java
@@ -1,0 +1,31 @@
+package org.programmers.springorder.dto.customer;
+
+import org.programmers.springorder.model.customer.Customer;
+
+import java.util.UUID;
+
+public class CustomerResponseDto {
+
+    private final UUID customerId;
+    private final String name;
+    private final String customerType;
+
+    public CustomerResponseDto(Customer customer) {
+        this.customerId = customer.getCustomerId();
+        this.name = customer.getName();
+        this.customerType = customer.getCustomerType().name();
+    }
+
+    public static CustomerResponseDto of(Customer customer) {
+        return new CustomerResponseDto(customer);
+    }
+
+    @Override
+    public String toString() {
+        return "customerId : " + customerId + '\n' +
+                "name : " + name + '\n' +
+                "customerType : " + customerType + '\n' +
+                "==============================";
+
+    }
+}
