+++ src/main/java/org/programmers/springorder/model/customer/CustomerType.java
@@ -1,0 +1,5 @@
+package org.programmers.springorder.model.customer;
+
+public enum CustomerType {
+    NORMAL, BLACK
+}
