+++ src/main/java/org/programmers/springorder/repository/voucher/MemoryVoucherRepository.java
@@ -1,0 +1,34 @@
+package org.programmers.springorder.repository.voucher;
+
+import org.programmers.springorder.model.voucher.Voucher;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Repository
+@Profile("local")
+public class MemoryVoucherRepository implements VoucherRepository {
+    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
+
+    @Override
+    public UUID save(Voucher voucher) {
+        storage.put(voucher.getVoucherId(), voucher);
+        return voucher.getVoucherId();
+    }
+
+    @Override
+    public List<Voucher> findAll() {
+        return storage.values().stream().toList();
+    }
+
+    @Override
+    public Optional<Voucher> findById(UUID voucherId) {
+        return Optional.ofNullable(storage.get(voucherId));
+    }
+
+}
