+++ src/main/java/org/programmers/springorder/service/CustomerService.java
@@ -1,0 +1,36 @@
+package org.programmers.springorder.service;
+
+import org.programmers.springorder.dto.customer.CustomerRequestDto;
+import org.programmers.springorder.dto.customer.CustomerResponseDto;
+import org.programmers.springorder.model.customer.Customer;
+import org.programmers.springorder.repository.customer.CustomerRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.UUID;
+
+@Service
+public class CustomerService {
+
+    private final CustomerRepository customerRepository;
+
+    public CustomerService(CustomerRepository customerRepository) {
+        this.customerRepository = customerRepository;
+    }
+
+    @Transactional
+    public Customer createCustomer(CustomerRequestDto customerRequestDto) {
+        Customer customer = Customer.of(UUID.randomUUID(), customerRequestDto);
+        customerRepository.save(customer);
+        return customer;
+    }
+
+    @Transactional
+    public List<CustomerResponseDto> getBlackList() {
+        return customerRepository.findAllBlackList()
+                .stream()
+                .map(CustomerResponseDto::of)
+                .toList();
+    }
+}
